Third commit
This commit is aimed at using the pyGTK library. It's the first time
I’ve ever played with it so it's not solid. Running the rssStart.py is
the way to run the program from command line without any GUI. Run
guyguiclass.py form any Debian installation or any computer that
already has the pygtk and python installed to run the program with a
GUI.